Move config into internal
