+++ internal/api/server.go
@@ -27,8 +27,8 @@ import (
 	"github.com/go-chi/httplog"
 	"github.com/rs/cors"
 	"github.com/stashapp/stash/internal/manager"
+	"github.com/stashapp/stash/internal/manager/config"
 	"github.com/stashapp/stash/pkg/logger"
-	"github.com/stashapp/stash/pkg/manager/config"
 	"github.com/stashapp/stash/pkg/models"
 	"github.com/stashapp/stash/pkg/utils"
 	"github.com/stashapp/stash/ui"
