+++ internal/api/session.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 
 	"github.com/stashapp/stash/internal/manager"
-	"github.com/stashapp/stash/pkg/manager/config"
+	"github.com/stashapp/stash/internal/manager/config"
 	"github.com/stashapp/stash/pkg/session"
 )
