+++ internal/manager/checksum.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 
+	"github.com/stashapp/stash/internal/manager/config"
 	"github.com/stashapp/stash/pkg/logger"
-	"github.com/stashapp/stash/pkg/manager/config"
 	"github.com/stashapp/stash/pkg/models"
 )
