+++ internal/manager/task_scan_gallery.go
@@ -8,10 +8,10 @@ import (
 	"strings"
 
 	"github.com/remeh/sizedwaitgroup"
+	"github.com/stashapp/stash/internal/manager/config"
 	"github.com/stashapp/stash/pkg/file"
 	"github.com/stashapp/stash/pkg/gallery"
 	"github.com/stashapp/stash/pkg/logger"
-	"github.com/stashapp/stash/pkg/manager/config"
 	"github.com/stashapp/stash/pkg/models"
 )
